+++ types.tsx
@@ -5,6 +5,12 @@
 
 import { MaterialIcons } from "@expo/vector-icons";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
+import {
+  GestureResponderEvent,
+  ImageSourcePropType,
+  StyleProp,
+  ViewStyle,
+} from "react-native";
 
 declare global {
   namespace ReactNavigation {
@@ -166,4 +172,28 @@ export type Sound = {
   hz: number;
   uri: string;
   lastModified: Date;
+};
+
+export type ButtonProps = {
+  onPress: (event: GestureResponderEvent) => void;
+  text: string;
+  outlined?: boolean;
+  danger?: boolean;
+  loading?: boolean;
+  disabled?: boolean;
+  style?: StyleProp<ViewStyle>;
+};
+
+export type Buttons = {
+  seeResult: ButtonProps;
+  newTest: ButtonProps;
+  sendTest: ButtonProps;
+};
+
+export type Page = {
+  title: string;
+  image: ImageSourcePropType;
+  subTitle: string;
+  description: string;
+  buttons: ButtonProps[];
 };
